Update searchbar to make it behave more like a launchbar application
