+++ stockflux-components/src/index.js
@@ -1,6 +1,8 @@
 import Titlebar from './components/Titlebar';
 
 import './styles/button-icon.css';
+import './styles/font.css';
+import './styles/scrollbar.css';
 import './index.css'
 
 export default {
