+++ stockflux-components/src/styles/font.css
@@ -1,0 +1,9 @@
+@import url(https://fonts.googleapis.com/css?family=Roboto:400,100,300,500);
+
+body, input {
+    font-family: "Roboto", sans-serif;
+    font-size: var(--font-normal);
+    font-weight: var(--font-weight-normal);
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+}
