+++ stockflux-search/package.json
@@ -13,6 +13,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-scripts": "^3.0.0",
+    "react-select": "^2.4.3",
     "stockflux-components": "file:../stockflux-components",
     "stockflux-core": "file:../stockflux-core",
     "throat": "^4.1.0"
