+++ stockflux-search/public/app.dev.json
@@ -1,4 +1,5 @@
 {
+    "devtools_port": 9091,
     "runtime": {
         "arguments": "--enable-aggressive-domstorage-flushing",
         "version": "10.66.39.43"
@@ -9,20 +10,22 @@
         "url": "http://localhost:3000/index.html",
         "applicationIcon": "http://localhost:3000/favicon.ico",
         "autoShow": true,
-        "defaultWidth": 230,
+        "defaultWidth": 325,
         "defaultHeight": 300,
-        "maxWidth": 230,
-        "minWidth": 230,
-        "minHeight": 300,
+        "resizable": false,
         "frame": false,
-        "contextMenu": true
+        "contextMenu": true,
+        "cornerRounding": {
+            "height": 4,
+            "width": 4
+        }
     },
     "services": [
         {
             "name": "layouts",
             "config": {
                 "features": {
-                    "tab": false
+                    "tab": true
                 }
             }
         }
