--- stockflux-search/src/components/Search.module.css
@@ -1,46 +1,0 @@
-.search {
-  --header-height: 45px;
-
-  width: var(--expanded-width);
-  height: 100%;
-  background-color: var(--search-background);
-  transition: all 0.2s linear 0.1s;
-  position: absolute;
-  box-shadow: 0 2px 0 var(--dark-background);
-}
-
-.header {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0 10px;
-  height: var(--header-height);
-}
-
-.searchInput {
-  border: none;
-  background: transparent;
-  font-size: var(--font-normal);
-  font-weight: var(--font-medium-light);
-  color: var(--text-grey);
-  flex: 1;
-  outline: none;
-}
-
-.searchIcon {
-  --icon-size: 20px;
-  width: var(--icon-size);
-  height: var(--icon-size);
-}
-
-.message {
-  padding: 10px;
-  text-align: center;
-  color: var(--text-grey);
-  font-size: var(--font-large);
-  font-weight: var(--font-medium);
-}
-
-.searchScroll {
-  height: calc(100% - var(--header-height) - var(--titlebar-height));
-}
