--- stockflux-search/src/components/SearchResult.module.css
@@ -1,39 +1,0 @@
-.searchResult {
-  height: var(--search-result-height);
-  border-bottom: 1px solid var(--separator-colour);
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.searchResult:hover {
-  background: var(--dark-background);
-}
-
-.star {
-  background-position: 0 50% !important;
-  transition: var(--icon-transition-duration) all;
-  cursor: pointer;
-}
-
-.name {
-  font-size: var(--font-normal);
-  font-weight: var(--font-thin);
-  line-height: 1.5;
-  width: 155px;
-  height: auto;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.code {
-  font-size: var(--font-xxxlarge);
-  font-weight: var(--font-medium);
-  line-height: 1.4;
-}
-
-.details {
-  padding-left: 20px;
-  flex: 1;
-}
