+++ stockflux-search/src/index.css
@@ -1,0 +1,13 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    height: 100%;
+    margin: 0;
+    background-color: var(--light-background);
+}
+
+#root {
+    height: 100%;
+}
