+++ stockflux-search/src/index.js
@@ -6,6 +6,8 @@ import 'malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar';
 import 'malihu-custom-scrollbar-plugin/jquery.mCustomScrollbar.css';
 import 'jquery-mousewheel/jquery.mousewheel';
 
+import './index.css';
+
 const mountApp = () => {
   ReactDOM.render(<App />, document.getElementById('root'));
 };
