+++ stockflux-search/src/search-result/index.js
@@ -1,0 +1,1 @@
+export { default }  from './SearchResult';
